+++ index.ts
@@ -1,3 +1,5 @@
 export { BlockEmitter } from "./src/BlockEmitter.js";
 export { createDefaultTransport } from "./src/createDefaultTransport.js";
+export { createNodeTransport } from "./src/createNodeTransport.js";
+export { createWebTransport } from "./src/createWebTransport.js";
 export { createHeadersInterceptor } from "./src/createHeadersInterceptor.js";
