+++ src/createDefaultTransport.ts
@@ -1,20 +1,4 @@
-import { createHeadersInterceptor } from "./createHeadersInterceptor.js";
-import type { IMessageTypeRegistry } from "@bufbuild/protobuf";
-import { Transport } from "@connectrpc/connect";
-import { createGrpcWebTransport } from "@connectrpc/connect-node";
-import { createAuthInterceptor } from "@substreams/core";
+import { createNodeTransport } from "./createNodeTransport.js";
 
-export function createDefaultTransport(baseUrl: string, token: string, registry: IMessageTypeRegistry, headers?: Headers): Transport {
-  const interceptors = [createHeadersInterceptor(headers)];
-  if (token) {
-    interceptors.push(createAuthInterceptor(token));
-  }
-  return createGrpcWebTransport({
-    baseUrl,
-    httpVersion: "2",
-    interceptors,
-    jsonOptions: {
-      typeRegistry: registry,
-    },
-  });
-}
+// make it backwards compatible for those still using `createDefaultTransport`
+export const createDefaultTransport = createNodeTransport;
